Run 6.22 miles
Another hot day. I ran a mile with Anna at about a 10:30 pace so it
brought my average pace down quite a bit. Nose breathing got pretty
tough on the last few miles. I learned to breath harder
through my nose and faster. I'm not used to it but it seemed to
help. It makes a satisfying noise.