Run 10.11 miles
This is probably my last long run in Ames. It was a nice. I went from
8:30pm to 10:00pm. Weather was perfect. So much better than running in
the sun. I'm going to miss these trails. Went all the way to the west
end and the back through campus town and brookside. Fireflies were
lighting the way and the sunset on the midwestern sky was
spectacular. It's been nice, Iowa. Leaving for the west coast with
on Friday!