Run 6.21 miles
I forgot to make this entry yesterday. I fixed Wednesday's run
entry which had the wrong timestamp.

The run was good. It was my second in Eugene so I made it a bit
farther on the trails. Still hot. The run back up the hill was brutal
again but this time I made it only on my nose. My legs were and still
are sore. I think they forgot how to run up hills from all that time
in Iowa.

I'm taking today off to get some rest.