Run 6.79 miles
I ditched my Nike Free's for a pair of Altra's today. They feel pretty
good. I went to the Run Hub in Eugene. They scanned my feet and told
me to get a pair of size 14's instead of 13. It was a great
experience. It feels good to be running more minimal again. I'm hoping
that the pain that I had when I was wearing sandals doesn't come
back. I learned from the scan that I put more weight on my left foot
than the right.  I bet its from all the BBall that I used to play.
I'm also in the 94 percentile of arch height.

Anyways - I'm heading back to california tomorrow. Hopefully the
running will be good. There are lots of fires there so it might
be smoky.