Add missing running entries
Both of these were up the logging road next to my mom's house in
litte river with Luna. I ran the first miles with Anna (~8:50's).
Running with the dog is kind of tough because of the leash. She likes
to pull me. She also takes lots of small stops to smell things.