Run 4.93 miles
Got hot in Eugene again. On Sunday I did a 8 mile hike up the south
sister with Anna and Christine. I think my legs were sort of sore for
this one. I cut it a bit short because we had early dinner plans, but
I was happy for the excuse.