Add RSS feed to posts
This works with ox-rss.el. The trick was generating a different
sitemap that is used for converting into RSS. Lots of help came from
here: https://writepermission.com/org-blogging-rss-feed.html

There is a bunch of commented out elisp in site.el that I used for
debugging. Will remove it in the next commit, but might be useful to
me in the future still.