Run 11.19 miles
Went up Comptche Ukiah Road. It was raining and I was really cold by
the end. I tracked the run with my Garmin GPS watch and Strava. The
garmin missed 2 miles! I've always thought that it might be missing
data because sometimes the pace seems way off, espcially when there
isn't a good line of site to the sky. Strava mapped my route perfectly
though, which is cool.